Update default delta_t2 value to be 9
Related: 
@cjyetman would you be in favour of removing this `determine_tdm_variables()` function?

To me this is setting configuration values (like the value of delta_t1 and delta_t2 values), which I feel like would make more sense to isolate in `workflow.transition.monitor`. 

Happy for the function to have sensible defaults, but I think we we are specifying the function configuration we should do it in a workflow.